compiler: Do not always force rebuild of compiler.

+++ build.sh
@@ -1,6 +1,12 @@
+#!/bin/bash
 
-rm -rf build
-mkdir build
+if [ "$1" == "--clean" ]; then
+    echo "Deleting previous SPECT compiler build..."
+    rm -rf build
+fi
+mkdir -p build
+
 cd build
 cmake ..
 make
+cd ..
